backend/pyproject.toml
[project]
name = "rag-backend"
version = "1.0.0"
description = "RAG Integration Backend for Physical AI Textbook"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "qdrant-client>=1.11.0",
    "openai>=1.54.0",
    "openai-agents>=0.1.0",
    "google-generativeai>=0.8.0",
    "asyncpg>=0.29.0",
    "slowapi>=0.1.9",
    "langchain>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "python-dotenv>=1.0.0",
    "pydantic[email]>=2.9.0",
    "pydantic-settings>=2.5.0",
    "passlib[argon2]>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.46",
    "psycopg2-binary>=2.9.11",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "black>=24.8.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["rag_backend"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
